endpoint/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pymongo==4.5.0
pytest==7.4.2
httpx==0.25.0
python-dotenv==1.0.0
